requirements.txt
alembic==1.8.1
asgiref==3.5.2
attrs==22.1.0
certifi==2022.9.24
charset-normalizer==2.1.1
click==8.1.3
distlib==0.3.6
Django==4.1.4
django-cors-headers==3.13.0
djangorestframework==3.14.0
djangorestframework-simplejwt==5.2.2
dnspython==2.2.1
dominate==2.7.0
exceptiongroup==1.0.3
filelock==3.8.0
Flask==2.2.2
Flask-Bootstrap==3.3.7.1
Flask-Login==0.6.2
Flask-Migrate==4.0.0
Flask-SQLAlchemy==3.0.2
Flask-WTF==1.0.1
greenlet==2.0.1
gunicorn==20.1.0
idna==3.4
iniconfig==1.1.1
itsdangerous==2.1.2
Jinja2==3.1.2
Mako==1.2.4
MarkupSafe==2.1.1
mysqlclient==2.1.1
numpy==1.23.4
packaging==21.3
pandas==1.5.1
Pillow==9.3.0
platformdirs==2.5.4
pluggy==1.0.0
PyJWT==2.6.0
pyparsing==3.0.9
pytest==7.2.0
python-dateutil==2.8.2
python-dotenv==0.21.0
pytz==2022.5
requests==2.28.1
six==1.16.0
SQLAlchemy==1.4.44
sqlparse==0.4.3
tomli==2.0.1
urllib3==1.26.12
virtualenv==20.16.7
visitor==0.1.3
Werkzeug==2.2.2
WTForms==3.0.1
